✨📡 dynamically convert between types of sources
This adds a new API to AtJSON, removing the `toCommonSchema` paradigm.
Instead, use `convertTo` to convert from one source to another.

Converters are defined on the class-level, using `defineConverterTo`
which takes the class to convert to, and a function that will convert
the document from one source to another.
