+++ packages/@atjson/source-gdocs-paste/src/source.ts
@@ -1,0 +1,28 @@
+import Document from '@atjson/document';
+import {
+  Bold,
+  Heading,
+  HorizontalRule,
+  Italic,
+  Link,
+  List,
+  ListItem,
+  Strikethrough,
+  Underline,
+  VerticalAdjust
+} from './annotations';
+
+import GDocsParser, { GDocsPasteBuffer } from './gdocs-parser';
+
+export default class extends Document {
+  static contentType = 'application/vnd.atjson+gdocs';
+  static schema = [Bold, Heading, HorizontalRule, Italic, Link, List, ListItem, Strikethrough, Underline, VerticalAdjust];
+  static fromSource(pasteBuffer: GDocsPasteBuffer) {
+    let gDocsParser = new GDocsParser(pasteBuffer);
+
+    return new this({
+      content: gDocsParser.getContent(),
+      annotations: gDocsParser.getAnnotations()
+    });
+  }
+}
